Simplify CircleCI
Before we have a fully functional system where all tests pass, there is
no point in investing so much in the CI pipeline which is prone to
failure. This change uses a single build step, which should be fast
(since its go) and contain most of the info we need.